rename directory to floatingChecklist
