+++ frontend/src/component/layout/MainLayout/MainLayout.tsx
@@ -15,8 +15,8 @@ import { ThemeMode } from 'component/common/ThemeMode/ThemeMode';
 import { NavigationSidebar } from './NavigationSidebar/NavigationSidebar.tsx';
 import { EventTimelineProvider } from 'component/events/EventTimeline/EventTimelineProvider';
 import { NewInUnleash } from './NavigationSidebar/NewInUnleash/NewInUnleash.tsx';
-import { FloatingOnboardingChecklist } from 'component/onboarding/floating/FloatingOnboardingChecklist.tsx';
-import { FloatingOnboardingChecklistProvider } from 'component/onboarding/floating/FloatingOnboardingChecklistContext.tsx';
+import { FloatingOnboardingChecklist } from 'component/onboarding/floatingChecklist/FloatingOnboardingChecklist.tsx';
+import { FloatingOnboardingChecklistProvider } from 'component/onboarding/floatingChecklist/FloatingOnboardingChecklistContext.tsx';
 import { AccessRequestsNotifications } from 'component/admin/users/AccessRequestsNotifications/AccessRequestsNotifications';
 import { useUiFlag } from 'hooks/useUiFlag';
