+++ frontend/src/component/menu/Header/HelpResources/HelpResources.tsx
@@ -20,8 +20,8 @@ import NewReleasesOutlinedIcon from '@mui/icons-material/NewReleasesOutlined';
 import RocketLaunchOutlinedIcon from '@mui/icons-material/RocketLaunchOutlined';
 import ExploreOutlinedIcon from '@mui/icons-material/ExploreOutlined';
 import LearningLabIcon from 'assets/icons/menu/learning-lab.svg?react';
-import { FloatingOnboardingChecklistContext } from 'component/onboarding/floating/FloatingOnboardingChecklistContext.tsx';
-import { OnboardingProgressBadge } from 'component/onboarding/floating/OnboardingProgressBadge.tsx';
+import { FloatingOnboardingChecklistContext } from 'component/onboarding/floatingChecklist/FloatingOnboardingChecklistContext.tsx';
+import { OnboardingProgressBadge } from 'component/onboarding/floatingChecklist/OnboardingProgressBadge.tsx';
 import { Link } from 'react-router';
 import { useFeedback } from 'component/feedbackNew/useFeedback';
 import { useEventTracker } from 'hooks/useEventTracker';
