+++ frontend/src/component/onboarding/floatingChecklist/ChecklistSteps.tsx
