+++ frontend/src/component/onboarding/floatingChecklist/FloatingOnboardingChecklist.tsx
