+++ frontend/src/component/onboarding/floatingChecklist/OnboardingProgressBadge.tsx
