+++ frontend/src/component/onboarding/floatingChecklist/floatingOnboardingChecklistState.ts
