+++ frontend/src/component/onboarding/floatingChecklist/useFirstProjectFeature.test.tsx
