+++ frontend/src/component/onboarding/floatingChecklist/useOnboardingChecklistEligibility.test.tsx
