+++ frontend/src/component/onboarding/floatingChecklist/useOnboardingChecklistEligibility.ts
